Refactor pfcwd_helpers.py
